# Sprite data reorg