--- a/src/game_logic/entity_factory.cpp
+++ b/src/game_logic/entity_factory.cpp
@@ -98,51 +98,53 @@ EntityFactory::EntityFactory(
 
 
 Sprite EntityFactory::createSpriteForId(const ActorID actorID) {
-  const auto actorParts = actorIDListForActor(actorID);
-  auto sprite = makeSpriteFromActorIDs(actorID, actorParts);
+  auto sprite = createSpriteComponent(actorID);
   configureSprite(sprite, actorID);
   return sprite;
 }
 
 
-const engine::OwningTexture& EntityFactory::getOrCreateTexture(
-  const IdAndFrameNr& textureId
-) {
-  auto it = mTextureCache.find(textureId);
-  if (it == mTextureCache.end()) {
-    const auto& actorData = mpSpritePackage->loadActor(textureId.first);
-    const auto& frameData = actorData.mFrames[textureId.second];
-    auto texture = engine::OwningTexture(
-      mpRenderer, frameData.mFrameImage);
-    it = mTextureCache.emplace(textureId, std::move(texture)).first;
-  }
-  return it->second;
-}
+Sprite EntityFactory::createSpriteComponent(const ActorID mainId) {
+  auto iData = mSpriteDataCache.find(mainId);
+  if (iData == mSpriteDataCache.end()) {
+    engine::SpriteDrawData drawData;
 
+    int lastDrawOrder = 0;
+    int lastFrameCount = 0;
+    std::vector<int> framesToRender;
 
-Sprite EntityFactory::makeSpriteFromActorIDs(
-  data::ActorID mainId,
-  const vector<ActorID>& actorIDs
-) {
-  Sprite sprite;
-  int lastFrameCount = 0;
-  int lastDrawOrder = 0;
+    const auto actorParts = actorIDListForActor(mainId);
+    for (const auto part : actorParts) {
+        const auto& actorData = mpSpritePackage->loadActor(part);
+        lastDrawOrder = actorData.mDrawIndex;
 
-  for (const auto ID : actorIDs) {
-    const auto& actorData = mpSpritePackage->loadActor(ID);
-    for (auto i=0u; i<actorData.mFrames.size(); ++i) {
-      const auto& frameData = actorData.mFrames[i];
-      const auto& texture = getOrCreateTexture(make_pair(ID, i));
-      const auto textureRef = engine::NonOwningTexture(texture);
-      sprite.mFrames.emplace_back(textureRef, frameData.mDrawOffset);
+        for (const auto& frameData : actorData.mFrames) {
+          auto texture = engine::OwningTexture{
+            mpRenderer, frameData.mFrameImage};
+          drawData.mFrames.emplace_back(
+            std::move(texture), frameData.mDrawOffset);
+        }
+
+        framesToRender.push_back(lastFrameCount);
+        lastFrameCount = int(actorData.mFrames.size());
     }
-    lastDrawOrder = actorData.mDrawIndex;
-    sprite.mFramesToRender.push_back(lastFrameCount);
-    lastFrameCount += static_cast<int>(actorData.mFrames.size());
+
+    drawData.mDrawOrder = adjustedDrawOrder(mainId, lastDrawOrder);
+
+    if (mainId == 5 || mainId == 6) {
+      for (int i=0; i<39; ++i) {
+        drawData.mFrames[i].mDrawOffset.x -= 1;
+      }
+    }
+
+    iData = mSpriteDataCache.emplace(
+      mainId,
+      SpriteData{std::move(drawData), std::move(framesToRender)}
+    ).first;
   }
 
-  sprite.mDrawOrder = adjustedDrawOrder(mainId, lastDrawOrder);
-  return sprite;
+  const auto& data = iData->second;
+  return {&data.mDrawData, data.mInitialFramesToRender};
 }
 
 
@@ -155,7 +157,7 @@ entityx::Entity EntityFactory::createSprite(
   entity.assign<Sprite>(sprite);
 
   if (assignBoundingBox) {
-    entity.assign<BoundingBox>(engine::inferBoundingBox(sprite.mFrames[0]));
+    entity.assign<BoundingBox>(engine::inferBoundingBox(sprite.mpDrawData->mFrames[0]));
   }
   return entity;
 }
@@ -179,7 +181,7 @@ entityx::Entity EntityFactory::createProjectile(
   auto entity = mpEntityManager->create();
   auto sprite = createSpriteForId(actorIdForProjectile(type, direction));
 
-  const auto boundingBox = engine::inferBoundingBox(sprite.mFrames[0]);
+  const auto boundingBox = engine::inferBoundingBox(sprite.mpDrawData->mFrames[0]);
 
   // TODO: Player projectiles shouldn't move on the frame they were created
   entity.assign<Active>();
@@ -198,7 +200,7 @@ entityx::Entity EntityFactory::createActor(
 ) {
   auto entity = createSprite(id, position);
   auto& sprite = *entity.component<Sprite>();
-  const auto boundingBox = engine::inferBoundingBox(sprite.mFrames[0]);
+  const auto boundingBox = engine::inferBoundingBox(sprite.mpDrawData->mFrames[0]);
 
   configureEntity(entity, id, boundingBox);
 
@@ -296,7 +298,7 @@ entityx::Entity EntityFactory::createEntitiesForLevel(
       boundingBox.topLeft = {0, 0};
     } else if (hasAssociatedSprite(actor.mID)) {
       const auto sprite = createSpriteForId(actor.mID);
-      boundingBox = engine::inferBoundingBox(sprite.mFrames[0]);
+      boundingBox = engine::inferBoundingBox(sprite.mpDrawData->mFrames[0]);
       entity.assign<Sprite>(sprite);
     }