Performance and memory usage optimization. Data structures for sprites are reorganized to only include a single pointer to the required data (`SpriteDrawData`), instead of a `vector` of `SpriteFrame` objects which have pointers to the textures. The list of `SpriteFrames` is the same for all sprites based on the same actor ID, but previously, each instance of a sprite would contain its own copy of that information. By getting rid of this redundant information, `sizeof(Sprite)` goes down from `72` to `48`, plus all the memory required for the `SpriteFrame` lists isn't needed anymore. 

More importantly, this reorganization allows caching the data on a coarser granularity, by mapping from actor ID to `SpriteDrawData*`, therefore reducing the amount of work required to create a sprite in the `EntityFactory` class. Previously, only the textures were cached, but each time a sprite was created, its frame data had to be loaded and processed, which means creating a sprite was an `O(N)` operation. Now, it's constant time after the first time a sprite for a particular ID was created. 

The change improves level load times for L1 on medium difficulty like follows:

| System | Compiler | Before | After | Speed gain |
| -------- | ------ | ------- | -------- | ------ |
| Desktop (Windows) | MSVC 2015 | 150 ms | 19 ms | 789 % |
| Desktop (Linnux) | Clang 3.9 | 23 ms | 8 ms | 288 % |
| Laptop (OS X) | Apple Clang 8.0.0 | 54 ms | 22 ms | 245 % |
| Raspberry Pi 1 (Model B) | GCC 5.4 | 1074 ms | 550 ms | 195 % |

Systems used for testing:

| Name | CPU | Memory |
| ------ | ---- | -------- |
| Desktop | Core i7 4770, 3.4 GHz | 16 GB DDR3 |
| Laptop | Core i5 4308U, 2.8 GHz | 8 GB DDR3 |
| Raspberry Pi | ARM 1176/BCM2835, 700 MHz | 512 MB |

Closes #147